234_02_003/work/requirements.txt
contourpy==1.2.1
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.53.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
openpyxl==3.1.3
packaging==24.0
pandas==2.0.3
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
seaborn==0.13.2
six==1.16.0
tzdata==2024.1
